Add support for Python3
This is a quick fix that allows to use make_parallel with Python3 while
still remaining compatible with Python2.

Division (/) was replaced with explicit integer division (//)
and xrange was replaced with range.
Since xrange is only used over GPU IDs and output layers of a model
there's no reason to prefer a generator over a list.